PERLIN LANDSCAPE

+++ Assets/Scripts/Graph/Node.cs
@@ -6,7 +6,7 @@ public class Node {
 	public string description;
     public int x, y;
 	public GameObject sceneObject;
-    public int height;
+    public float height;
 
 	public Node(int _x, int _y, GameObject o = null) {
         this.x = _x;
